pyproject.toml
[project]
name = "adk-agent-shopping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "crewai-tools>=0.45.0",
    "google-adk[eval]>=1.0.0",
    "langchain-community>=0.3.24",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "ruff>=0.11.10",
    "youtube-search>=2.1.2",
    "youtube-transcript-api>=1.0.3",
]

[tool.black]
line-length = 119
exclude = "venv"

[tool.flake8]
docstring-convention = "all"
ignore = ["D107", "D212", "E501", "W503", "W605", "D203", "D100"]
exclude = ["venv"]
max-line-length = 119

# radon
radon-max-cc = 10

[tool.isort]
profile = "black"
multi_line_output = "VERTICAL_HANGING_INDENT"
force_grid_wrap = 2
line_length = 119

[tool.pylint."messages control"]
disable = [
    "line-too-long",
    "trailing-whitespace",
    "missing-function-docstring",
    "consider-using-f-string",
    "import-error",
    "too-few-public-methods",
    "redefined-outer-name",
]

[tool.ruff.lint]
select = ["F", "E", "W", "I", "UP"]
ignore = ["E501", "W292"]
unfixable = ["ERA001"]


[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401", "F403"]
"**/{tests,docs,tools}/*" = ["E402"]
